extract toolbar cleanup logic to helper class
