+++ src/Form/Extension/ToolbarFormExtension.php
@@ -9,6 +9,7 @@
 
 namespace App\Form\Extension;
 
+use App\Form\Helper\ToolbarHelper;
 use App\Form\Toolbar\ExportToolbarForm;
 use App\Form\Toolbar\InvoiceToolbarForm;
 use App\Form\Toolbar\InvoiceToolbarSimpleForm;
@@ -18,22 +19,16 @@ use App\Form\Toolbar\UserToolbarForm;
 use App\Reporting\MonthlyUserListForm;
 use App\Reporting\WeeklyUserListForm;
 use App\Reporting\YearlyUserListForm;
-use App\User\TeamService;
-use App\User\UserService;
 use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\FormBuilderInterface;
 
 final class ToolbarFormExtension extends AbstractTypeExtension
 {
-    private $userService;
-    private $teamService;
-    private $teamNames = ['team', 'teams', 'searchTeams'];
-    private $userNames = ['user', 'users'];
+    private $toolbarHelper;
 
-    public function __construct(UserService $userService, TeamService $teamService)
+    public function __construct(ToolbarHelper $toolbarHelper)
     {
-        $this->userService = $userService;
-        $this->teamService = $teamService;
+        $this->toolbarHelper = $toolbarHelper;
     }
 
     public static function getExtendedTypes(): iterable
@@ -53,36 +48,6 @@ final class ToolbarFormExtension extends AbstractTypeExtension
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $deleteUser = false;
-        foreach ($this->userNames as $name) {
-            if ($builder->has($name) && $this->userService->countUser(true) < 2) {
-                $deleteUser = true;
-                break;
-            }
-        }
-
-        if ($deleteUser) {
-            foreach ($this->userNames as $name) {
-                if ($builder->has($name)) {
-                    $builder->remove($name);
-                }
-            }
-        }
-
-        $deleteTeams = false;
-        foreach ($this->teamNames as $name) {
-            if ($builder->has($name) && !$this->teamService->hasTeams()) {
-                $deleteTeams = true;
-                break;
-            }
-        }
-
-        if ($deleteTeams) {
-            foreach ($this->teamNames as $name) {
-                if ($builder->has($name)) {
-                    $builder->remove($name);
-                }
-            }
-        }
+        $this->toolbarHelper->cleanupForm($builder);
     }
 }
